Docs/add swagger docs

+++ src/main/java/com/elibrary/group4/controller/Admin/BookController.java
@@ -1,5 +1,7 @@
 package com.elibrary.group4.controller.Admin;
 
+import com.elibrary.group4.Utils.Validation.JwtUtil;
+import com.elibrary.group4.exception.ForbiddenException;
 import com.elibrary.group4.model.Book;
 import com.elibrary.group4.model.request.BookRequest;
 import com.elibrary.group4.model.response.SuccessResponse;
@@ -22,8 +24,15 @@ public class BookController {
     @Autowired
     ModelMapper modelMapper;
 
+    @Autowired
+    JwtUtil jwtUtil;
     @PostMapping
-    public ResponseEntity createBook(@Valid BookRequest request) throws Exception {
+    public ResponseEntity createBook(@RequestHeader("Authorization") String token, @Valid BookRequest request) throws Exception {
+        var tokenAndRole = jwtUtil.getRoleAndId(token);
+
+        if (!tokenAndRole.get("role").equals("USER")) {
+            throw new ForbiddenException("Forbidden");
+        }
         Book book =  bookService.create(request);
         return ResponseEntity.status(HttpStatus.CREATED).body(new SuccessResponse<Book>("Created",book));
     }
@@ -33,21 +42,38 @@ public class BookController {
             @RequestParam(defaultValue = "1") Integer page,
             @RequestParam(defaultValue = "5") Integer size,
             @RequestParam(defaultValue = "DESC") String direction,
-            @RequestParam(defaultValue = "bookId") String sortBy
+            @RequestParam(defaultValue = "bookId") String sortBy,
+            @RequestParam(defaultValue = "") String name,
+            @RequestParam(defaultValue = "") String author,
+            @RequestParam(defaultValue = "") String language,
+            @RequestParam(defaultValue = "") String releaseYear,
+            @RequestParam(defaultValue = "") String category
     ) throws Exception {
-        Page<Book> books =bookService.list(page, size, direction, sortBy);
+        Page<Book> books =bookService.listBooksUsingSpecification(page, size, sortBy, direction, name, author, releaseYear, language, category);
         return ResponseEntity.status(HttpStatus.OK).body(new SuccessResponse<>("Success",books));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity update(@Valid @RequestBody BookRequest request, @PathVariable("id") String id) throws Exception{
+    public ResponseEntity update(@RequestHeader(name = "Authorization") String token, @Valid @RequestBody BookRequest request, @PathVariable("id") String id) throws Exception{
+        var tokenAndRole = jwtUtil.getRoleAndId(token);
+
+        if (!tokenAndRole.get("role").equals("ADMIN")) {
+            throw new ForbiddenException("Forbidden");
+        }
         bookService.update(request,id);
         return ResponseEntity.status(HttpStatus.OK).body(new SuccessResponse<>("Updated", request));
     }
 
 
     @DeleteMapping("/{id}")
-    public ResponseEntity delete(@PathVariable("id")String id) throws Exception{
+    public ResponseEntity delete(@RequestHeader(name = "Authorization") String token, @PathVariable("id")String id) throws Exception{
+
+        var tokenAndRole = jwtUtil.getRoleAndId(token);
+
+        if (!tokenAndRole.get("role").equals("ADMIN")) {
+            throw new ForbiddenException("Forbidden");
+        }
+
         bookService.delete(id);
         return ResponseEntity.status(HttpStatus.OK).body(new SuccessResponse<>("Deleted",null));
     }
